requirements_vercel.txt
Flask==2.3.3
Werkzeug==2.3.7
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
mysql-connector-python==8.1.0
reportlab==4.0.4
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
Authlib==1.2.1
